pyproject.toml
[project]
name = "realtime-cc-fraud-detector"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.121.1",
    "google-cloud-storage>=3.5.0",
    "numpy>=2.0.2",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "tensorflow==2.18.0",
    "uvicorn>=0.38.0",
]
